+++ src/app/components/drawer/Projects/SettingsComponent.js
@@ -15,6 +15,7 @@ const messages = defineMessages({
     defaultMessage: 'Annotations',
     description: 'Label for the Annotations settings navigation menu item',
   },
+  // to be deleted
   data: {
     id: 'teamSettingsNavigation.data',
     defaultMessage: 'Data',
@@ -116,6 +117,7 @@ const SettingsComponent = ({
               </li>
             </Link> : null
           }
+          {/* to be deleted */}
           { isAdminOrEditor && Boolean(team.smooch_bot) ?
             <Link className={cx('team-settings__data-tab', styles.linkList)} title={intl.formatMessage(messages.data)} to={`/${team.slug}/settings/data`}>
               <li className={cx([styles.listItem], { [styles.listItem_active]: tab === 'data' })}>
