--- src/app/components/drawer/Projects/TiplineDashboard.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const TiplineDashboard = () => (
-  <div>TiplineDashboard</div>
-);
-
-// eslint-disable-next-line import/no-unused-modules
-export default TiplineDashboard;
