+++ src/app/components/team/TeamComponent.js
@@ -11,13 +11,13 @@ import TeamRules from './Rules';
 import TeamStatuses from './Statuses';
 import SmoochBot from './SmoochBot';
 import TeamTags from './TeamTags';
-import TeamData from './TeamData';
 import TeamTasks from './TeamTasks';
 import TeamReport from './TeamReport';
 import TeamMembers from './TeamMembers';
 import TeamIntegrations from './TeamIntegrations';
 import TeamSimilarity from './Similarity';
 import Newsletter from './Newsletter';
+import TeamData from '../dashboard';
 import PageTitle from '../PageTitle';
 import UserUtil from '../user/UserUtil';
 import CheckContext from '../../CheckContext';
